Fixes a wrong project import and removes telerik reference.
The module project did not load in visual studio due to the project file trying to load MsBuild.Community.Tasks before a nuget restore happened.

Although this module was already telerik free, there was still a project reference to telerik.web.ui.dll

This does not really affect the build and is just a cleanup for next release.